version 2 of the dockerization of DataCatalogue

+++ src/main/java/com/admir/demiraj/datacatalogspringboot/service/VariablesXLSX_JSON.java
@@ -46,12 +46,14 @@ public class VariablesXLSX_JSON
     private VersionDAO versionDAO;
     @Autowired
     private StorageService storageService;
-    private String PATHOLOGY_CODE="";//we keepin the code of the pathology here (we might use it in the concept-paths)
+    public String PATHOLOGY_CODE="";//we keepin the code of the pathology here (we might use it in the concept-paths)
 
     public String filePath = "";
     public Versions version = null;
     public Versions harmonizedVersion = null;
     public Hospitals hospital = null;
+    // If we are on edit mode we shouldn't delete a version in case of an error
+    public boolean editMode = false;
 
     /**
      * @param file: the path of the input XLSX file
@@ -171,6 +173,8 @@ public class VariablesXLSX_JSON
         Iterator<Variables> it = xlsxVars.iterator();
         while (it.hasNext())
             addPathNodes(it.next(), root);
+
+        System.out.println("Finished create tree");
         return root;
     }
     /**
@@ -223,9 +227,10 @@ public class VariablesXLSX_JSON
     private void addPathNodes(Variables nextVar, Node root)
     {
         String thisConceptPath = nextVar.getConceptPath();
+        System.out.println("code is"+nextVar.getCode()+"add path notes concept path"+thisConceptPath);
 
 
-        System.out.println("variable name: "+nextVar.getName()+" variable concept path: "+nextVar.getConceptPath());
+        //System.out.println("variable name: "+nextVar.getName()+" variable concept path: "+nextVar.getConceptPath());
         if (thisConceptPath==null || thisConceptPath.trim().equals("") || thisConceptPath.trim().equals("/") /*|| thisConceptPath.trim().equals("/root") || thisConceptPath.trim().equals("/root/")*/)
         {
             //thisConceptPath=root.getName()+nextVar.getCode();
@@ -246,7 +251,7 @@ public class VariablesXLSX_JSON
             root.setCode(conceptPath[0]);
             //root.setConceptPath(conceptPath[0]);//------ ***** TO-DO: CHECK DATABASE IF THERE IS SUCH A PATHOLOGY <-------- *********
             PATHOLOGY_CODE = conceptPath[0];//we store the code of the pathology to have it for later...
-            System.out.println("~~~~~~******* We are talking about " + PATHOLOGY_CODE + "*******~~~~~~");
+            System.out.println("~~~~~~******* We are talking about " + PATHOLOGY_CODE + "*******~~~~~~ concept path: "+thisConceptPath);
             if (thisConceptPath.startsWith("/root/"))
             {
                 thisConceptPath.replaceFirst("/root/", "/"+PATHOLOGY_CODE+"/");
@@ -695,11 +700,14 @@ public class VariablesXLSX_JSON
 
 public void throwExceptioAndDelete(String exceMessage,String exceDetails, String exceNextSteps){
     storageService.moveFileToErrorFiles(this.filePath);
-    if(this.hospital==null){
+    // We should delete versions with errors only when we are not on edit mode
+    if(this.hospital==null && !editMode){
         // delete cde version
+        System.out.println("Deleting CDEVersion");
         versionDAO.deleteVersion(this.version);
-    }else {
+    }else if(!editMode){
         // delete variables version
+        System.out.println("Deleting variable version both normal and harmonized");
         versionDAO.deleteVersion(this.hospital,this.version);
         versionDAO.deleteVersion(this.hospital,this.harmonizedVersion);
     }
